+++ tsup.config.ts
@@ -17,4 +17,5 @@ export default defineConfig({
   banner: {
     js: '#!/usr/bin/env node',
   },
+  external: ['readline/promises'],
 });